Code update

+++ Configuration.h
@@ -17,28 +17,28 @@
 
 
 @interface Configuration : NSObject {
-
-	NSString *server;
-	NSString *tag;
-	NSString *logfile;
-	NSString *protocol;
-	NSString *periodicity;
-	NSString *ocsPkgFilePath;
-	NSString *modulesFilePath;
-	NSString *serverdirFilePath;
-	NSString *cacertFilePath;	
-	NSString *launchdFilePath;
-	NSString *nowFilePath;
-	NSString *cfgFilePath;
-	
-	int debugmode;
-	int download;
-	int lazy;
-	int startup;
-	int now;
-
-	NSFileManager *filemgr;
-	
+    
+    NSString *server;
+    NSString *tag;
+    NSString *logfile;
+    NSString *protocol;
+    NSString *periodicity;
+    NSString *ocsPkgFilePath;
+    NSString *modulesFilePath;
+    NSString *serverdirFilePath;
+    NSString *cacertFilePath;
+    NSString *launchdFilePath;
+    NSString *nowFilePath;
+    NSString *cfgFilePath;
+    
+    NSInteger debugmode;
+    NSInteger download;
+    NSInteger lazy;
+    NSInteger startup;
+    NSInteger now;
+    
+    NSFileManager *filemgr;
+    
 }
 
 //Accessors for variables
@@ -63,20 +63,20 @@
 - (NSString *) cacertFilePath;
 - (void) setCacertFilePath:(NSString *)newCacertFilePath;
 
-- (int) debugmode;
-- (void) setDebugmode:(int)newDebugmode;
+- (NSInteger) debugmode;
+- (void) setDebugmode:(NSInteger)newDebugmode;
 
-- (int) download;
-- (void) setDownload:(int)newDownload;
+- (NSInteger) download;
+- (void) setDownload:(NSInteger)newDownload;
 
-- (int) lazy;
-- (void) setLazy:(int)newLazy;
+- (NSInteger) lazy;
+- (void) setLazy:(NSInteger)newLazy;
 
-- (int) startup;
-- (void) setStartup:(int)newStartup;
+- (NSInteger) startup;
+- (void) setStartup:(NSInteger)newStartup;
 
-- (int) now;
-- (void) setNow:(int)newNow;
+- (NSInteger) now;
+- (void) setNow:(NSInteger)newNow;
 
 
 @end
